pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nerve"
version = "0.0.0"
description = "NERVE: Neuro-symbolic Evidence Review and Verification Engine for LLM bio-agents."
authors = [{ name = "Biohackathon Team" }]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
  "pyyaml>=6.0.0",
  "streamlit>=1.40.0",
  "gliner2>=0.1.0",
  "transformers>=4.35.0",
  "neo4j>=6.0.3",
  "scikit-learn>=1.7.2",
  "pyvis>=0.3.2",
]

[dependency-groups]
dev = [
  "ruff~=0.6.0",
  "mypy~=1.10.0",
  "pytest~=8.3.0",
  "black>=25.11.0",
  "types-PyYAML>=6.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
strict = true
mypy_path = ["src", "stubs"]
exclude = ["^scripts/"]

[tool.black]
line-length = 100
target-version = ['py313']

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end tests that make network requests (deselect with '-m not e2e')",
]
addopts = "-m 'not e2e'"
